feat: update ci files in regard to cypress selection
