+++ package.json
@@ -35,6 +35,7 @@
     "fs-extra": "^8.1.0",
     "handlebars": "4.5.3",
     "inquirer": "^7.0.1",
+    "js-yaml": "3.13.1",
     "mkdirp": "^0.5.1"
   },
   "repository": {
@@ -51,6 +52,7 @@
     "@types/handlebars": "4.1.0",
     "@types/inquirer": "6.5.0",
     "@types/jest": "25.1.2",
+    "@types/js-yaml": "3.12.2",
     "@types/mkdirp": "1.0.0",
     "@typescript-eslint/eslint-plugin": "2.20.0",
     "@typescript-eslint/parser": "2.20.0",
