+++ src/services/index.ts
@@ -1,0 +1,7 @@
+export * from 'services/copyAssets';
+export * from 'services/updatePackageJson';
+export * from 'services/updateCiFiles';
+export * from 'services/updateStoreConfig';
+export * from 'services/installDependencies';
+export * from 'services/gererateComponent';
+export * from 'services/generateStore';
